the utility file iteration gets a method to report high time watermark
the inbound feed abstraction gets it's save/load state mechanism.
it serializes to YAML in var/state/infeeds/FEEDNAME.state

all the fake rigging that forced it to yield changes in the last day now
instead rolls the state forward via a commit method to preserve knowledge
we're "done" with the processed content, by rolling the cursor timestamp to
the highest seen in the process.

the abstract infeed class now gets the whole feed config, not just it's
specfific config, as it's useful. the local config is now in @localconfig

not much else... still spitting out filenames only, but it will now
only spit out the NEW ones. :)

still much much more to do. ;)