The infeed source for teamsite now emits RawContentItems vs filenames
It's currently a bit hardcoded to some radio records, but the mappings
will be externalized to the feed config file to allow adjustment
to whatever...

I'll probably also later push up the general "new XML files appearing" pipe
logic into an abstract that could be equally useful for NewsML processing.
(Just different xpath to property mappings... and subsequent transform rules)

One of the other next steps is a generalized RSS feed input adapter following the
same abstract interface. That will complete converson of the "top row" of the
Endeca pipeline in terms of conversion of existing logic.

Much progress. Much more to do. But for now... sleep!