a new Baron::InboundFeed::Transformer, which takes rawcontent objects, and runs mapping rules on them. It will support many, but for the moment only a single rule action "copy" is implemented, which copies a raw field into a native content attribute. Some introspection is done to avoid a big case block for action methods...
the run inbound script in 'bin' has been setup to test this, and now dumps
both the raw object, and a native for each item.

So, we now emit consistent (if sparsely defined...) native content objects
whether the source was rss or teamsite. Only title and date are currently
defined, but the mechanism is general and all we need now is:

- implement the native ContentItem object to spec.
- enhance our mappings
- add new rules
- implement new rules where we need them (AS generalized operators please!)