Main feature here is a generic "application frame", so the various apps will have things like a standardized log mechanism, config file, top level exception handling, etc...
The inbound feeds script in bin/ is now obsolete, the new way to
run the same thing is:

 /opt/local/bin/ruby bin/run-app.rb infeeds www_flare_com

Really just have the basics of instantiation, and logging setup,
much to do. ;)

The other thing done here is that this work was the first exploration
into splitting up the Baron module into various files, now I've a handle
on how that works... (like PHP requires). Now that what I've tried seems
to work, I'll split out the rest. The convention will be that modules will
result in both a file, and a directory made in their name. Sub-modules will
do the same within their parent directory, and classes will be files in
their module's directory. Every file should have it's external dependancies
declared entirely via requires.

Phew... a mouthfull for a wiki page...