Quick experiment on changing the 'mapper' thing with the Transform object up for an eval() instead.... probably the way we'll go, as maintenance of the command block will be a PITA.
For the moment, it will work with either a ruleEval, or the old way, preferring the first if
both are there... will probably drop the transformer before long... or it will change to a
utility class of functions to support the eval block.

Context of ruleEval is that "item" and "raw" are preset, raw is "full", and item is "empty".
By the end of the block, item must be filled.